pyproject.toml
[tool.poetry]
name = "Natural D"
version = "0.1.0"
description = ""
authors = ["Hemanth Nandipati <maaki@duck.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sentiment-nltk = "^0.0.2"
joblib = "^1.3.2"
scikit-learn = "1.0.2"
praw = "^7.7.1"
tweepy = "^4.14.0"
sqlalchemy = "^2.0.21"
xgboost = "^2.0.0"
flask = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
